implemented the error handler and error class in multiple api endpoints
