+++ src/app/api/course-list/route.ts
@@ -1,20 +1,15 @@
 import { NextResponse } from "next/server";
-import { connectToDatabase } from "@/lib/database/mongoose";
-import { Course } from "@/db/course";
+import { getCourses } from "@/lib/services/subjects";
+import { customErrorHandler } from "@/lib/utils/error";
 
 export const dynamic = "force-dynamic";
 
 export async function GET() {
   try {
-    await connectToDatabase();
-    const courses = await Course.find().lean();
+    const courses = await getCourses();
 
     return NextResponse.json(courses, { status: 200 });
   } catch (error) {
-    console.error(error);
-    return NextResponse.json(
-      { message: "Failed to fetch courses", error },
-      { status: 500 },
-    );
+    return customErrorHandler(error, "Failed to fetch courses");
   }
 }
