+++ src/app/api/related-subject/route.ts
@@ -1,41 +1,22 @@
 import { NextResponse, type NextRequest } from "next/server";
-import { connectToDatabase } from "@/lib/database/mongoose";
-import { IRelatedSubject } from "@/interface";
-import RelatedSubject from "@/db/relatedSubjects";
-import { escapeRegExp } from "@/lib/utils/regex";
+import { customErrorHandler } from "@/lib/utils/error";
+import { getRelatedSubjects } from "@/lib/services/subjects";
 
 export const dynamic = "force-dynamic";
 
 export async function GET(req: NextRequest) {
   try {
-    await connectToDatabase();
-    const url = req.nextUrl.searchParams;
-    const subject = url.get("subject");
-
-    if (!subject) {
-      return NextResponse.json(
-        { message: "Subject query parameter is required" },
-        { status: 400 },
-      );
-    }
-
-    const escapedSubject = escapeRegExp(subject);
-    const subjects: IRelatedSubject[] = await RelatedSubject.find({
-      subject: { $regex: new RegExp(`${escapedSubject}`, "i") },
-    });
     
-    console.log("realted", subjects);
+    const url = req.nextUrl.searchParams;
+    const sub = url.get("subject");
+
+    const subjects = await getRelatedSubjects(sub || "");
 
     return NextResponse.json(
-      {
-        related_subjects: subjects[0]?.related_subjects
-      },
+      { related_subjects: subjects[0]?.related_subjects },
       { status: 200 },
     );
   } catch (error) {
-    return NextResponse.json(
-      { message: "Failed to fetch related subject", error },
-      { status: 500 },
-    );
+    return customErrorHandler(error, "Failed to fetch related subject");
   }
 }
