+++ src/lib/services/paper.ts
@@ -3,10 +3,12 @@ import { type IPaper } from "@/interface";
 import { escapeRegExp } from "@/lib/utils/regex";
 import { extractUniqueValues } from "@/lib/utils/paper-aggregation";
 import { connectToDatabase } from "../database/mongoose";
+import { CustomError } from "@/lib/utils/error";
+import { Types } from "mongoose";
 
 export async function getPapersBySubject(subject: string) {
     if (!subject){
-        throw new Error("Subject query parameter is required");
+        throw new CustomError("Subject query parameter is required", 400);
     }
 
     await connectToDatabase();
@@ -23,4 +25,19 @@ export async function getPapersBySubject(subject: string) {
         ...uniqueValues,
     }
 
+}
+
+export async function getPapersById(id: string) {
+    if (!Types.ObjectId.isValid(id)) {
+        throw new CustomError("Invalid paper ID", 400);
+    }
+    await connectToDatabase();
+
+    const paper = await Paper.findById(id);
+
+    if (!paper) {
+        throw new CustomError("Paper not found", 404);
+    }
+
+    return paper;
 }
